[MiniMax-M3] clean up vLLM sparse attention adapter
Keep MiniMax-M3 sparse attention on the dedicated adapter path and rename the backend metadata pieces to match their model-specific ownership.
